requirements.txt
asgiref==3.8.1
attrs==25.1.0
cffi==1.17.1
constantly==23.10.4
cryptography==44.0.2
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.7.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.0
msgpack==1.1.0
packaging==24.2
psycopg==3.2.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==25.0.0
python-dotenv==1.0.1
redis==5.2.0
service-identity==24.2.0
setuptools==75.8.2
sqlparse==0.5.3
typing_extensions==4.12.2
python-socketio==5.11.1
python-engineio==4.9.0
eventlet==0.35.1